requirements.txt
beautifulsoup4==4.8.1
neo4j-driver==1.7.5
neobolt==1.7.15
neotime==1.7.4
pytz==2019.3
six==1.12.0
soupsieve==1.9.4
urllib3==1.25.6
